requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
huggingface-hub==0.16.4
openai==1.3.7
python-dotenv==1.0.0
click==8.1.7
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
joblib==1.3.2
transformers==4.35.2
torch==2.1.1 